Added a helper message for showing a message that deals with the timeout and makes sure the timeout only removes the message that was just shown, not any future messages.
Although it's a good idea anyway to have a helper, it's mainly to correct a bug.  If you for example delete a lot of code (which has a timeout of 7 seconds on the message), then close the banner, and overall 6.5 seconds later delete more code, that message will be hidden after only 0.5 seconds because the original timeout fires and closes whatever message is showing.  So we need to keep track to make sure we only hide the message after 7 seconds if it is still the same message showing and not a new one.  To track this, we add a new field to the store that increments each time we show a new message with content.