Change how we disable print() and input() when doing autocomplete.
Instead of looking for the code and replacing it (which wasn't working when the input() was on the line we were trying to get AC for), we configure Skulpt to do nothing on print() and return "" on input(), then leave the code as-is.  This is simpler and seems to work fine.

Fixes #242