Changed the selection tests to use Playwright instead of Cypress.
Cypress was not able to do shift+left/right, and even the complicated test rig could not emulate the behaviour in the browser (which deletes text if you select across slots and type).  Playwright seems like it's actually an accurate test (plus, across three browsers).  The tests do not yet pass but I think they might be accurate at least.